Wave-1 adversarial reviews: districts, simulation, reporting, analytics/export
- districts: exact-fraction apportionment ties, stricter validation (PlanValidation
  is a list of errors), merge-split recombination (60 splits, PP median 0.286)
- simulation: shared presidential ticket shocks across provinces, NaN guards,
  scenario include hardening; demo-2028 retuned (contingent rate 38% -> 12%)
- reporting: national PRES race excluded from calling, manual-call replay,
  checkpointed rewinds, realistic Dutch count timing
- analytics/export: contest-level selection by race type, no double counting of
  presidential totals, lossless integer/datetime export, streaming writers

+++ src/app/districts/plan.py
@@ -41,7 +41,7 @@ class GeneratedPlan:
     unit_province: np.ndarray  #: (U,) object — province code of each unit
     unit_municipality: np.ndarray  #: (U,) object — CBS municipality code of each unit
     unit_population: np.ndarray  #: (U,) int64
-    unit_district: np.ndarray  #: (U,) int32 — index into the district arrays
+    unit_district: np.ndarray  #: (U,) int64 — index into the district arrays
     unit_overridden: np.ndarray  #: (U,) bool — assignment set by a manual override
     # --- districts (D) ---------------------------------------------------------------------
     district_codes: list[str]
